Sign added, a check for having visited a screen, and empty states for the rest of the beginning
Added the sign screen which switches back and forth between itself and the open reef state, and a check to make sure that the text boc doesn't keep repeating itself whenever you start openReefState. I also added a genericState if we want to copy and paste to make the art of creating a state ultra-easy. Just substitute in the various generic words for what you need to use. Also created a few empty states I'm gonna get to work on immediately.